requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2021.5.30        # via requests
charset-normalizer==2.0.5  # via requests
idna==3.2                 # via requests
numpy==1.21.2             # via pandas
pandas==1.3.3             # via -r requirements.in
pydantic==1.8.2           # via -r requirements.in
python-dateutil==2.8.2    # via pandas
python-dotenv==0.19.0     # via -r requirements.in
pytz==2021.1              # via pandas
pyyaml==5.4.1             # via -r requirements.in
requests==2.26.0          # via -r requirements.in
six==1.16.0               # via python-dateutil
typing-extensions==3.10.0.2  # via pydantic
urllib3==1.26.6           # via requests
